Fix MongoDB host name and remove unnecessary code
